requirements.txt
requests>=2.31.0
schedule>=1.2.1
supabase>=2.3.0
python-dotenv>=1.0.0
tabulate>=0.9.0 